move sdk data to correct components, instead of all in dev credentials
