+++ src/context/DevCredentialsContext.tsx
@@ -23,17 +23,8 @@ interface DevCredentialsContextProps {
   clientId?: string;
   apiKey?: string;
   redirectUri?: string;
-  permissionTemplateId?: string;
-  setCredentials: (credentials: {
-    clientId: string;
-    apiKey: string;
-    redirectUri: string;
-  }) => void;
   credentialsLoading: boolean; // Renamed to avoid conflict with AuthContext
   invalidCredentials: boolean;
-  vehicleTokenIds?: string[];
-  vehicleMakes?: string[];
-  transactionData?: TransactionData;
   uiState: string;
   setUiState: React.Dispatch<React.SetStateAction<string>>;
   componentData: any;
@@ -51,22 +42,11 @@ export const DevCredentialsProvider = ({
   children: ReactNode;
 }): JSX.Element => {
   const [clientId, setClientId] = useState<string | undefined>();
-  const [uiState, setUiState] = useState("EMAIL_INPUT"); //TODO: should be enum
   const [apiKey, setApiKey] = useState<string | undefined>();
   const [redirectUri, setRedirectUri] = useState<string | undefined>();
-  const [permissionTemplateId, setPermissionTemplateId] = useState<
-    string | undefined
-  >();
-  const [vehicleTokenIds, setVehicleTokenIds] = useState<
-    string[] | undefined
-  >();
-  const [vehicleMakes, setVehicleMakes] = useState<string[] | undefined>();
-  const [transactionData, setTransactionData] = useState<
-    TransactionData | undefined
-  >();
   const [credentialsLoading, setCredentialsLoading] = useState<boolean>(true); // Renamed loading state
   const [invalidCredentials, setInvalidCredentials] = useState<boolean>(false);
-
+  const [uiState, setUiState] = useState("EMAIL_INPUT"); //TODO: should be enum
   const [componentData, setComponentData] = useState<any | null>(null);
 
   // Example of using postMessage to receive credentials (as described previously)
@@ -76,42 +56,12 @@ export const DevCredentialsProvider = ({
 
     const clientIdFromUrl = urlParams.get("clientId");
     const redirectUriFromUrl = urlParams.get("redirectUri");
-    const permissionTemplateIdFromUrl = urlParams.get("permissionTemplateId");
-    const vehiclesArrayFromUrl = urlParams.getAll("vehicles");
-    const vehiclesMakesArrayFromUrl = urlParams.getAll("vehicleMakes");
-    const transactionDataFromUrl = urlParams.get("transactionData");
     const entryStateFromUrl = urlParams.get("entryState");
 
     if (clientIdFromUrl && redirectUriFromUrl) {
       setClientId(clientIdFromUrl);
       setRedirectUri(redirectUriFromUrl);
       setApiKey("api key"); //not needed for redirect url
-
-      //Optional Param Checks
-      if (permissionTemplateIdFromUrl != null) {
-        // Checks for both null and undefined
-        setPermissionTemplateId(permissionTemplateIdFromUrl);
-      }
-
-      if (vehiclesArrayFromUrl != null) {
-        setVehicleTokenIds(vehiclesArrayFromUrl);
-      }
-
-      if (vehiclesMakesArrayFromUrl != null) {
-        setVehicleTokenIds(vehiclesMakesArrayFromUrl);
-      }
-
-      if (transactionDataFromUrl != null) {
-        try {
-          const parsedTransactionData = JSON.parse(
-            decodeURIComponent(transactionDataFromUrl)
-          );
-
-          setTransactionData(parsedTransactionData);
-        } catch (error) {
-          console.error("Failed to parse transactionData:", error);
-        }
-      }
 
       setUiState(entryStateFromUrl || "EMAIL_INPUT");
 
@@ -122,21 +72,13 @@ export const DevCredentialsProvider = ({
           eventType,
           clientId,
           apiKey,
-          permissionTemplateId,
           redirectUri,
-          vehicles,
           entryState,
-          vehicleMakes,
-          transactionData,
         } = event.data;
         if (eventType === "AUTH_INIT") {
           setClientId(clientId);
           setApiKey(apiKey || "api key"); //todo, bring back when api key is needed
           setRedirectUri(redirectUri);
-          setPermissionTemplateId(permissionTemplateId);
-          setVehicleTokenIds(vehicles);
-          setVehicleMakes(vehicleMakes);
-          setTransactionData(transactionData);
           setUiState(entryState || "EMAIL_INPUT");
           setCredentialsLoading(false); // Credentials loaded
         }
@@ -188,13 +130,8 @@ export const DevCredentialsProvider = ({
         clientId,
         apiKey,
         redirectUri,
-        permissionTemplateId,
-        setCredentials,
         credentialsLoading,
         invalidCredentials,
-        vehicleTokenIds,
-        vehicleMakes,
-        transactionData,
         uiState,
         setUiState,
         componentData,
